+++ svd-rs/src/field.rs
@@ -44,18 +44,8 @@ impl Field {
 #[cfg(feature = "serde")]
 mod ser_de {
     use super::*;
+    use crate::{DeserArray, SerArray};
     use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-    #[derive(serde::Deserialize, serde::Serialize)]
-    struct FieldArray {
-        #[cfg_attr(
-            feature = "serde",
-            serde(flatten, default, skip_serializing_if = "Option::is_none")
-        )]
-        dim: Option<DimElement>,
-        #[cfg_attr(feature = "serde", serde(flatten))]
-        info: FieldInfo,
-    }
 
     impl Serialize for Field {
         fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -64,11 +54,7 @@ mod ser_de {
         {
             match self {
                 Self::Single(info) => info.serialize(serializer),
-                Self::Array(info, dim) => FieldArray {
-                    dim: Some(dim.clone()),
-                    info: info.clone(),
-                }
-                .serialize(serializer),
+                Self::Array(info, dim) => SerArray { dim, info }.serialize(serializer),
             }
         }
     }
@@ -78,7 +64,7 @@ mod ser_de {
         where
             D: Deserializer<'de>,
         {
-            let FieldArray { dim, info } = FieldArray::deserialize(deserializer)?;
+            let DeserArray { dim, info } = DeserArray::<FieldInfo>::deserialize(deserializer)?;
             if let Some(dim) = dim {
                 Ok(info.array(dim))
             } else {